requirements.txt
certifi>=2024.0.0
aiogram==3.27.0
aiohttp==3.13.5
aiosqlite==0.22.1
APScheduler==3.11.2
beautifulsoup4==4.14.3
lxml==6.1.0
python-dotenv==1.2.2
tenacity==9.1.4
pytest==9.0.3
pytest-asyncio==1.3.0
